requirements.txt
bcrypt==3.1.7
blinker==1.4
certifi==2019.3.9
cffi==1.12.3
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mongoengine==0.18.2
Pillow==6.1.0
pycparser==2.19
pymongo==3.9.0
six==1.12.0
Werkzeug==0.16.0
WTForms==2.2.1
